requirements.txt
streamlit
pandas
plotly.express
openpyxl
nltk
seaborn
scikit-learn